Reorganized directories->easier project management
