+++ client/src/components/home/Slideshow.js
